The Origami Embedded Controller firmware (Origami-EC)
Origami-EC early release, version pre-square.

Note that infrastructure for the project is not ready at this point and
that no documentation, form of support or public statement will be
provided. This is a semi-public release.

On the other hand, the code is starting to look good and can conceivably
start receiving external contributions. However, future changes will
likely be merged into a single commit when the project is publicly
announced.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>