Introduce banked segments switching support
This introduces support for banked segments switching, also known as
bank switching or banking. It allows circumventing the 64 kiB code size
limit of the 8051 by dynamically remapping banks of memory to different
segments that may live outside of the 64 kiB limit.

Trampoline code for the KB9012 platform is provided, as well as default
definitions of device and platform-specific segments.

Note that SDCC currently only allows segments to reside 64 kiB above
their mapped code offset, with no way of controlling that offset.
This limits the switching ability to a single 4 kiB segment.

To demonstrate the banking capabilities, 2 distinct 2 kiB segments are
switched instead (but are not controlled separately).

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>