Complete banked segments switching support
This completes banked segments switching support by allowing the use of
the full flash. This was previously impossible due to SDCC defining the
call address (when executing in the same segment) as the lower 64 kiB of
the base address in memory, only allowing one per 64 kiB chunks of
memory and thus only allowing switching between 2 segments at most.

The trick put in place consists of defining the base addresses of
segments with the same lower 64 kiB, matching their run-time addresses.
Segments are later moved around in the flash image with the segments
tool. The bank switching assembly routine will then use the remaining
part of the address as an offset.

Signed-off-by: Paul Kocialkowski <contact@paulk.fr>